FUNCIONA
al parecer es proteus el que muestra titilante, no es mi programa

+++ LIBRERÍAS.X/Config_ADC.c
@@ -1,7 +1,7 @@
 #include "Config_ADC.h"
 
 int VAL;
-uint8_t VAL_SWAP;
+int VAL_SWAP;
 int VAL_NIBBLE1;
 int VAL_NIBBLE2;
 
@@ -11,17 +11,17 @@ int ADC(uint8_t ADRESL_, uint8_t ADRESH_){
     return VAL;
 }
 // FUNCION PARA HACER SWAP EL VALOR DE LA CONVERSION
-uint8_t SWAP_ADC(uint8_t VAL_ADC){
+int SWAP_ADC(uint8_t VAL_ADC){
     VAL_SWAP = (((VAL_ADC & 0x0F)<<4) | ((VAL_ADC & 0xF0)>>4));
     return VAL_SWAP;
 }
 
-uint8_t NIBBLE1_ADC(uint8_t VAL_ADC){
+int NIBBLE1_ADC(uint8_t VAL_ADC){
     VAL_NIBBLE1 = (VAL_ADC & 15);
     return VAL_NIBBLE1;
 }
 
-uint8_t NIBBLE2_ADC(uint8_t VAL_SWAP){
+int NIBBLE2_ADC(uint8_t VAL_SWAP){
     VAL_NIBBLE2 = (VAL_SWAP & 15);
     return VAL_NIBBLE2;
 }
