+++ LIBRERÍAS.X/Config_ADC.h
@@ -20,9 +20,9 @@
 // en el display y se carga los leds del display para que aparezca
 //******************************************************************************
 int ADC(uint8_t ADRESL_, uint8_t ADRESH_);
-uint8_t SWAP_ADC(uint8_t VAL_ADC);
-uint8_t NIBBLE1_ADC(uint8_t VAL_ADC);
-uint8_t NIBBLE2_ADC(uint8_t VAL_SWAP);
+int SWAP_ADC(uint8_t VAL_ADC);
+int NIBBLE1_ADC(uint8_t VAL_ADC);
+int NIBBLE2_ADC(uint8_t VAL_SWAP);
 
 #endif	/* CONFIG_ADC_H */
